Try better using pytest fixtures instead of _pre_setup or _post_teardown
