+++ apps/amo/tests/__init__.py
@@ -10,9 +10,8 @@ from datetime import datetime, timedelta
 from functools import partial, wraps
 from urlparse import parse_qs, urlparse, urlsplit, urlunsplit
 
-from django import forms, http, test
+from django import forms, test
 from django.conf import settings
-from django.core.cache import cache
 from django.core.management import call_command
 from django.db.models.signals import post_save
 from django.forms.fields import Field
@@ -20,7 +19,6 @@ from django.http import SimpleCookie
 from django.test.client import Client, RequestFactory
 from django.utils import translation
 
-import caching
 import mock
 import pytest
 import tower
@@ -45,7 +43,6 @@ from applications.models import AppVersion
 from bandwagon.models import Collection
 from files.models import File
 from lib.es.signals import process, reset
-from translations.hold import clean_translations
 from translations.models import Translation
 from versions.models import ApplicationsVersions, Version
 from users.models import RequestUser, UserProfile
@@ -257,82 +254,18 @@ class MockEsMixin(object):
                 Mocked_ES.stop()
 
 
-JINJA_INSTRUMENTED = False
-
-
-def default_prefixer():
-    """Make sure each test starts with a default URL prefixer."""
-    request = http.HttpRequest()
-    request.META['SCRIPT_NAME'] = ''
-    prefixer = amo.urlresolvers.Prefixer(request)
-    prefixer.app = settings.DEFAULT_APP
-    prefixer.locale = settings.LANGUAGE_CODE
-    amo.urlresolvers.set_url_prefix(prefixer)
-
-
-def test_pre_setup():
-    cache.clear()
-    # Override django-cache-machine caching.base.TIMEOUT because it's
-    # computed too early, before settings_test.py is imported.
-    caching.base.TIMEOUT = settings.CACHE_COUNT_TIMEOUT
-
-    translation.trans_real.deactivate()
-    # Django fails to clear this cache.
-    translation.trans_real._translations = {}
-    translation.trans_real.activate(settings.LANGUAGE_CODE)
-
-    # Make sure the "templates" list in a response is properly updated,
-    # even though we're using Jinja2 and not the default django template
-    # engine.
-    global JINJA_INSTRUMENTED
-    if not JINJA_INSTRUMENTED:
-        import jinja2
-        old_render = jinja2.Template.render
-
-        def instrumented_render(self, *args, **kwargs):
-            context = dict(*args, **kwargs)
-            test.signals.template_rendered.send(
-                sender=self, template=self, context=context)
-            return old_render(self, *args, **kwargs)
-
-        jinja2.Template.render = instrumented_render
-        JINJA_INSTRUMENTED = True
-
-    # Reset the prefixer.
-    default_prefixer()
-
-
-def test_post_teardown():
-    amo.set_user(None)
-    clean_translations(None)  # Make sure queued translations are removed.
-
-    # Make sure we revert everything we might have changed to prefixers.
-    amo.urlresolvers.clean_url_prefixes()
-
-
 class BaseTestCase(test.TestCase):
-    """Base test case that each and every test cases should inherit from."""
+    """Base test case that most test cases should inherit from."""
 
     def _pre_setup(self):
         super(BaseTestCase, self)._pre_setup()
-        test_pre_setup()
         self.client = self.client_class()
 
-    def _post_teardown(self):
-        test_post_teardown()
-        super(BaseTestCase, self)._post_teardown()
-
-    def shortDescription(self):
-        # Stop nose using the test docstring and instead the test method name.
-        pass
-
     def trans_eq(self, trans, localized_string, locale):
-        eq_(Translation.objects.get(id=trans.id,
-                                    locale=locale).localized_string,
-            localized_string)
+        translation = Translation.objects.get(id=trans.id, locale=locale)
+        assert translation.localized_string == localized_string
 
-    def assertUrlEqual(self, url, other, compare_host=False,
-                       compare_scheme=False):
+    def assertUrlEqual(self, url, other, compare_host=False):
         """Compare url paths and query strings."""
         assert_url_equal(url, other, compare_host=compare_host)
